+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Translations: [简体中文](CHANGELOG.zh.md)
 * change: Deprecate AsyncImage with keepContentNoneStartOnDraw parameter
 * change: Deprecate the fitScale property of TransitionTarget, adding TransitionViewTarget.scaleType
   and TransitionComposeTarget.contentsScale instead of fitScale
+* new: CrossfadePainter now supports contentScale
 
 # 4.1.0